refactor: delete ActionAnonymizer
