+++ great_expectations/core/usage_statistics/anonymizers/anonymizer.py
@@ -190,3 +190,17 @@ class Anonymizer:
             pass
 
         return None
+
+    def anonymize_action_info(self, action_name, action_obj=None, action_config=None):
+        anonymized_info_dict: dict = {
+            "anonymized_name": self.anonymize(action_name),
+        }
+
+        self.anonymize_object_info(
+            object_=action_obj,
+            object_config=action_config,
+            anonymized_info_dict=anonymized_info_dict,
+            runtime_environment={"module_name": "great_expectations.checkpoint"},
+        )
+
+        return anonymized_info_dict
